sick stuff

+++ src/system/char/CharacterTest.h
@@ -66,8 +66,8 @@ protected:
     /** "graphically displays the screensize and lod next to the character" */
     bool mShowScreenSize; // 0x8b
     bool mShowFootExtents; // 0x8c
-    int unk90;
-    int unk94;
+    float unk90; // 0x90 - transition beat threshold
+    int unk94; // 0x94
     ClipDistMap *unk98;
     RndOverlay *mOverlay; // 0x9c
 };
